gke-sample/pyproject.toml
[tool.poetry]
name = "gke-sample"
version = "0.1.0"
description = ""
authors = ["KK <ery.eden@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<3.9"
numpy = "^1.21.2"
pulumi-gcp = "^5.20.0"
pulumi-kubernetes = "^3.7.3"
google-cloud-storage = "^1.42.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
